refactor(timeblock): plan-service からバリデーション・ガードを分離する
plan-service.ts（443行）から validateRange / ensurePlanCanBeCreated /
assertOptimisticLock / isPastPlan / overlap ガード / エラー整形を
plan-guards.ts の関数群へ移動。ロジックは書き換えずに移動のみ、
PlanService の public API は不変。両ファイルとも 400 行未満。

Refs #1528
